feat(refactor): relocated query_slots to work_with_slots
