+++ backend/tools/for_db/work_with_slots.py
@@ -1,4 +1,4 @@
-from sqlalchemy import and_
+from sqlalchemy import and_, or_
 from db import models
 from db.models import Slots, SlotsShema
 
@@ -77,3 +77,17 @@ def get_slot_by_id(slots_id):
     slots_shema = SlotsShema(many=False)
     output = slots_shema.dump(id_added_slots)
     return output
+
+
+def query_slots(admin_id, date, end_interval, none_or_not=None):
+    list_of_slots = Slots.query.filter(
+        and_(
+            or_(Slots.start_interval.between(date, end_interval),
+                Slots.end_interval.between(date, end_interval),
+                and_(Slots.start_interval <= date, Slots.end_interval >= end_interval)),
+            or_(none_or_not is None,
+                none_or_not),
+            Slots.admin_id == admin_id
+        )
+    )
+    return list_of_slots
